chore: consolidate shared methods, update examples, clean up formdata headers
